refactor: unify auth forms
